+++ FirebaseAppCheck/Sources/Core/FIRAppCheckLogger.m
@@ -22,18 +22,12 @@ NS_ASSUME_NONNULL_BEGIN
 
 FIRLoggerService kFIRLoggerAppCheck = @"[FirebaseAppCheck]";
 
-NSString *const kFIRLoggerAppCheckMessageCodeUnknown = @"I-FAA001001";
-
 // FIRAppCheck.m
 NSString *const kFIRLoggerAppCheckMessageCodeProviderFactoryIsMissing = @"I-FAA002001";
 NSString *const kFIRLoggerAppCheckMessageCodeProviderIsMissing = @"I-FAA002002";
 
-// FIRAppCheckAPIService.m
-NSString *const kFIRLoggerAppCheckMessageCodeUnexpectedHTTPCode = @"I-FAA003001";
-
 // FIRAppCheckDebugProvider.m
 NSString *const kFIRLoggerAppCheckMessageDebugProviderIncompleteFIROptions = @"I-FAA004001";
-NSString *const kFIRLoggerAppCheckMessageDebugProviderFailedExchange = @"I-FAA004002";
 
 // FIRAppCheckDebugProviderFactory.m
 NSString *const kFIRLoggerAppCheckMessageCodeDebugToken = @"I-FAA005001";
@@ -41,16 +35,30 @@ NSString *const kFIRLoggerAppCheckMessageCodeDebugToken = @"I-FAA005001";
 // FIRDeviceCheckProvider.m
 NSString *const kFIRLoggerAppCheckMessageDeviceCheckProviderIncompleteFIROptions = @"I-FAA006001";
 
-// FIRAppAttestProvider.m
-NSString *const kFIRLoggerAppCheckMessageCodeAppAttestNotSupported = @"I-FAA007001";
-NSString *const kFIRLoggerAppCheckMessageCodeAttestationRejected = @"I-FAA007002";
-
 #pragma mark - Log functions
+
 void FIRAppCheckDebugLog(NSString *messageCode, NSString *message, ...) {
   va_list args_ptr;
   va_start(args_ptr, message);
   FIRLogBasic(FIRLoggerLevelDebug, kFIRLoggerAppCheck, messageCode, message, args_ptr);
   va_end(args_ptr);
+}
+
+#pragma mark - Helper functions
+
+GACAppCheckLogLevel FIRGetGACAppCheckLogLevel(void) {
+  FIRLoggerLevel loggerLevel = FIRGetLoggerLevel();
+  switch (loggerLevel) {
+    case FIRLoggerLevelError:
+      return GACAppCheckLogLevelError;
+    case FIRLoggerLevelWarning:
+    case FIRLoggerLevelNotice:
+      return GACAppCheckLogLevelWarning;
+    case FIRLoggerLevelInfo:
+      return GACAppCheckLogLevelInfo;
+    case FIRLoggerLevelDebug:
+      return GACAppCheckLogLevelDebug;
+  }
 }
 
 NS_ASSUME_NONNULL_END