+++ FirebaseInAppMessaging/Sources/Private/Data/FIRIAMMessageContentDataWithImageURL.h
@@ -38,7 +38,7 @@ NS_ASSUME_NONNULL_BEGIN
  *                   it's easier for doing mocking with unit testing.
  */
 - (instancetype)initWithMessageTitle:(nullable NSString *)title
-                         messageBody:(NSString *)body
+                         messageBody:(nullable NSString *)body
                     actionButtonText:(nullable NSString *)actionButtonText
            secondaryActionButtonText:(nullable NSString *)secondaryActionButtonText
                            actionURL:(nullable NSURL *)actionURL