+++ Firestore/firestore_collection_group_index_config.tf
@@ -1,0 +1,14 @@
+locals {
+  collection_group_indexes = {
+    index1 = [
+      {
+        field_path = "testId"
+        order      = "ASCENDING"
+      },
+      {
+        field_path = "a"
+        order      = "ASCENDING"
+      },
+    ]
+  }
+}