+++ ReleaseTooling/Sources/FirebaseManifest/Pod.swift
@@ -54,14 +54,4 @@ public struct Pod {
   public func podspecName() -> String {
     return "\(name).podspec"
   }
-
-  /// The Firebase pod does not support import validation with Xcode 12 because of the deprecated
-  /// ML pods not supporting the ARM Mac slice.
-  public func skipImportValidation() -> String {
-    if name == "Firebase" {
-      return "--skip-import-validation"
-    } else {
-      return ""
-    }
-  }
 }